Time: 184 ms (58.11%), Space: 67.5 MB (48.57%) - LeetHub

+++ 1679-max-number-of-k-sum-pairs/1679-max-number-of-k-sum-pairs.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int maxOperations(vector<int>& nums, int k) {
-        unordered_map<int,int> m; 
+        unordered_map<int,short int> m; 
         int n = nums.size();
         int ans = 0;
         for(int i = 0 ; i < n ; i++)
@@ -16,11 +16,6 @@ public:
                 m[k - nums[i]]++;
                 
             }
-            // else
-            // {
-            //     nums[i] = -1;
-            // }
-            
         }
         return ans;
     }
